split AI, add bounded addr tracking
